make it easier to measure startup time
Add an option to exit the daemon after probing. The magical invocation is

 # time ./run-hald.sh --skip-fdi-install --exit-after-probing > /dev/null 2>&1

running from the hald/ directory (need one normal run to install the
fdi files). On my Macbook Pro I get these figures

 real    0m2.084s
 user    0m0.564s
 sys     0m0.576s

+++ hald/run-hald.sh
@@ -3,13 +3,17 @@
 export HALD_RUNNER_PATH=`pwd`/linux:`pwd`/linux/probing:`pwd`/linux/addons:`pwd`/.:`pwd`/../tools:`pwd`/../tools/linux
 export PATH=`pwd`/../hald-runner:$PATH
 
-rm -rf .local-fdi
-make -C ../fdi install DESTDIR=`pwd`/.local-fdi prefix=/
+if [ "$1" = "--skip-fdi-install" ] ; then
+    shift
+else
+    rm -rf .local-fdi
+    make -C ../fdi install DESTDIR=`pwd`/.local-fdi prefix=/
+fi
 export HAL_FDI_SOURCE_PREPROBE=.local-fdi/share/hal/fdi/preprobe
 export HAL_FDI_SOURCE_INFORMATION=.local-fdi/share/hal/fdi/information
 export HAL_FDI_SOURCE_POLICY=.local-fdi/share/hal/fdi/policy
 
-./hald --daemon=no --verbose=yes
+./hald --daemon=no --verbose=yes $@
 #./hald --daemon=no
